S3US7_Securisation_des_routes - sécurisation de la route Skill/delete

+++ src/Controller/SkillController.php
@@ -96,6 +96,10 @@ class SkillController extends AbstractController
     #[Route('/{id}', name: 'app_skill_delete', methods: ['POST'])]
     public function delete(Request $request, Skill $skill, EntityManagerInterface $entityManager): Response
     {
+        if (!($this->security->isGranted('ROLE_COLLABORATEUR'))) {
+            return $this->redirectToRoute('app_home');
+        }
+
         if ($this->isCsrfTokenValid('delete' . $skill->getId(), $request->request->get('_token'))) {
             $entityManager->remove($skill);
             $entityManager->flush();
